_fave_nba_hoopers_refactored/requirements.txt
nba-api==1.5.2
pandas==2.2.0
numpy==1.26.3
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
